vmm: Use usize for IRQ type instead of u64
IRQs are mostly used as an index into an array or vector and should
be usize.  This requires some conversion to/from u64 at the interface
with seL4, but eliminates almost all conversion within the vmm.

+++ src/rust/components/vmm/config/src/lib.rs
@@ -7,7 +7,7 @@ use serde::{Serialize, Deserialize};
 use icecap_config_common::{DescTimerClient, DescMappedRingBuffer};
 use icecap_sel4_hack::prelude::*;
 
-pub type IRQ = u64;
+pub type IRQ = usize;
 
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct Config {
